Move PHI handling to SimpleStudyExportContext
