+++ study/api-src/org/labkey/api/study/writer/SimpleStudyExportContext.java
@@ -3,8 +3,10 @@ package org.labkey.api.study.writer;
 import org.jetbrains.annotations.Nullable;
 import org.labkey.api.admin.LoggerGetter;
 import org.labkey.api.data.Container;
+import org.labkey.api.data.PHI;
 import org.labkey.api.security.User;
 import org.labkey.api.specimen.Vial;
+import org.labkey.api.study.model.ParticipantMapper;
 import org.labkey.api.writer.VirtualFile;
 import org.labkey.study.xml.StudyDocument;
 
@@ -21,9 +23,16 @@ public class SimpleStudyExportContext extends AbstractContext
     private List<String> _participants = new ArrayList<>();
     private List<Vial> _vials = null;
 
-    protected SimpleStudyExportContext(User user, Container c, StudyDocument studyDoc, Set<String> dataTypes, LoggerGetter logger, @Nullable VirtualFile root)
+    private final PHI _phiLevel;
+    private final ParticipantMapper _participantMapper;
+    private final boolean _maskClinic;
+
+    protected SimpleStudyExportContext(User user, Container c, StudyDocument studyDoc, Set<String> dataTypes, PHI phiLevel, ParticipantMapper participantMapper, boolean maskClinic, LoggerGetter logger, @Nullable VirtualFile root)
     {
         super(user, c, studyDoc, dataTypes, logger, root);
+        _phiLevel = phiLevel;
+        _participantMapper = participantMapper;
+        _maskClinic = maskClinic;
     }
 
     public Set<Integer> getVisitIds()
@@ -54,5 +63,34 @@ public class SimpleStudyExportContext extends AbstractContext
     public void setVials(List<Vial> vials)
     {
         _vials = vials;
+    }
+
+    @Override
+    public PHI getPhiLevel()
+    {
+        return _phiLevel;
+    }
+
+    @Override
+    public boolean isAlternateIds()
+    {
+        return getParticipantMapper().isAlternateIds();
+    }
+
+    @Override
+    public boolean isMaskClinic()
+    {
+        return _maskClinic;
+    }
+
+    @Override
+    public boolean isShiftDates()
+    {
+        return getParticipantMapper().isShiftDates();
+    }
+
+    public ParticipantMapper getParticipantMapper()
+    {
+        return _participantMapper;
     }
 }
