+++ study/src/org/labkey/study/importer/CreateChildStudyPipelineJob.java
@@ -46,6 +46,7 @@ import org.labkey.api.study.Visit;
 import org.labkey.api.study.importer.SimpleStudyImporter;
 import org.labkey.api.study.importer.SimpleStudyImporterRegistry;
 import org.labkey.api.study.model.ParticipantGroup;
+import org.labkey.api.study.model.ParticipantMapper;
 import org.labkey.api.view.NotFoundException;
 import org.labkey.api.view.ViewContext;
 import org.labkey.api.writer.MemoryVirtualFile;
@@ -56,7 +57,6 @@ import org.labkey.study.model.ChildStudyDefinition;
 import org.labkey.study.model.DatasetDefinition;
 import org.labkey.study.model.ParticipantCategoryImpl;
 import org.labkey.study.model.ParticipantGroupManager;
-import org.labkey.study.model.ParticipantMapper;
 import org.labkey.study.model.StudyImpl;
 import org.labkey.study.model.StudyManager;
 import org.labkey.study.model.StudySnapshot;
